requirements.txt
aiofiles==24.1.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.2
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==4.9.4
MarkupSafe==3.0.2
plivo==4.56.1
priority==2.0.0
PyJWT==2.9.0
python-dotenv==1.0.1
Quart==0.19.6
requests==2.32.3
six==1.16.0
taskgroup==0.0.0a4
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.4
wsproto==1.2.0
mysql-connector-python>=8.0.33